requirements.txt
langgraph==0.0.48
openai==1.3.7
pandas==2.1.4
networkx==3.2.1
faiss-cpu==1.7.4
streamlit==1.28.1
python-dotenv==1.0.0
numpy==1.24.3